# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/workspace": {
      "get": {
        "operationId": "workspace",
        "parameters": [
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resolve a workspace by slug or tenant id",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/providers": {
      "get": {
        "operationId": "providers",
        "parameters": [
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "List auth providers for a tenant (secrets redacted)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/providers/admin": {
      "get": {
        "operationId": "providersAdmin",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin view of auth provider configs",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/providers/oidc": {
      "patch": {
        "operationId": "updateOidc",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOidcProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable/configure OIDC SSO for the tenant",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/providers/saml": {
      "patch": {
        "operationId": "updateSaml",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSamlProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable/configure SAML 2.0 SSO for the tenant",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/saml/metadata": {
      "get": {
        "operationId": "samlMetadata",
        "parameters": [
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "SP metadata for SAML IdP configuration",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/saml/start": {
      "get": {
        "operationId": "samlStart",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Begin SAML SP-initiated flow",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/saml/acs": {
      "get": {
        "operationId": "samlAcsGet",
        "parameters": [
          {
            "name": "RelayState",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "SAMLResponse",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "SAML ACS (mock/dev GET handler)",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "samlAcsPost",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "SAML ACS (HTTP-POST binding)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/oidc/start": {
      "get": {
        "operationId": "oidcStart",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Begin OIDC Authorization Code + PKCE flow",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/oidc/callback": {
      "get": {
        "operationId": "oidcCallback",
        "parameters": [
          {
            "name": "error",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OIDC callback — exchanges code and redirects to web",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Register first admin user for a tenant",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Create a session (bearer token)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/mfa/verify": {
      "post": {
        "operationId": "verifyMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaVerifyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Complete MFA challenge after password login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/mfa/setup": {
      "post": {
        "operationId": "startMfa",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start TOTP enrollment (returns secret + otpauth URL)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/mfa/confirm": {
      "post": {
        "operationId": "confirmMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm TOTP enrollment with an authenticator code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/mfa/disable": {
      "post": {
        "operationId": "disableMfa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDisableDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable TOTP with a current authenticator code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/invite/{token}": {
      "get": {
        "operationId": "getInvite",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Preview a user invite",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/invite/accept": {
      "post": {
        "operationId": "acceptInvite",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept invite and set password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password-reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Request password reset (enumeration-safe)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password-reset/{token}": {
      "get": {
        "operationId": "getPasswordReset",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Preview a password reset token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password-reset/confirm": {
      "post": {
        "operationId": "confirmPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Confirm password reset",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current session or API-key principal",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tenant users (admin)",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTenantUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create user with password (admin)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/users/invite": {
      "post": {
        "operationId": "inviteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite user without password (admin)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/delegations/candidates": {
      "get": {
        "operationId": "listDelegationCandidates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Users eligible as delegation targets",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Procure Ledger API",
    "description": "Phase 1 modular monolith — auth session, master data, invoices, capture/files, integration jobs. Authenticate with a session bearer token or a `pl_…` API key (legacy `aptora_…` keys still accepted).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT / session / pl_ API key",
        "type": "http",
        "description": "Session token from POST /api/auth/login (or invite accept), or an API key starting with pl_ (legacy aptora_ accepted)"
      }
    },
    "schemas": {
...
      "UpdateOidcProviderDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/OidcSettingsDto"
          }
        }
      },
...
      "UpdateSamlProviderDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/SamlSettingsDto"
          }
        }
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "totpCode": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "MfaConfirmDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  }
}